@@ -12,40 +12,29 @@
 		"format": "prettier --write ."
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-static": "^3.0.4",
-		"@sveltejs/kit": "^2.5.26",
+		"@sveltejs/adapter-static": "^3.0.5",
+		"@sveltejs/kit": "^2.5.28",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0-next.7",
 		"@types/eslint": "^9.6.1",
-		"@typescript-eslint/eslint-plugin": "^8.5.0",
-		"@typescript-eslint/parser": "^8.5.0",
+		"@typescript-eslint/eslint-plugin": "^8.7.0",
+		"@typescript-eslint/parser": "^8.7.0",
 		"autoprefixer": "^10.4.20",
-		"eslint": "^9.10.0",
+		"eslint": "^9.11.1",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.43.0",
-		"flowbite": "^2.5.1",
-		"flowbite-svelte": "^0.46.16",
-		"flowbite-svelte-icons": "^1.6.1",
-		"postcss": "^8.4.45",
+		"eslint-plugin-svelte": "^2.44.0",
+		"flowbite-svelte-icons": "^2.0.0-next.16",
+		"postcss": "^8.4.47",
 		"prettier": "^3.3.3",
 		"prettier-plugin-svelte": "^3.2.6",
 		"prettier-plugin-tailwindcss": "^0.6.6",
-		"svelte": "^5.0.0-next.244",
-		"svelte-check": "^4.0.1",
-		"tailwindcss": "^3.4.10",
+		"svelte": "^5.0.0-next.258",
+		"svelte-5-ui-lib": "^0.9.15",
+		"svelte-check": "^4.0.2",
+		"tailwindcss": "^3.4.13",
 		"tslib": "^2.7.0",
 		"typescript": "^5.6.2",
-		"vite": "^5.4.3",
+		"vite": "^5.4.7",
 		"vite-plugin-tailwind-purgecss": "^0.3.3"
 	},
-	"type": "module",
-	"overrides": {
-		"flowbite-svelte": {
-			"svelte": "^5.0.0"
-		},
-		"flowbite-svelte-icons": {
-			"svelte": "^5.0.0"
-		}
-	},
-	"dependencies": {
-		"@sveltejs/vite-plugin-svelte": "^4.0.0-next.7"
-	}
+	"type": "module"
 }
