@@ -1,68 +1,39 @@
 <script lang="ts">
   import { onNavigate } from '$app/navigation'
-  import { page } from '$app/stores'
-  import {
-    CloseButton,
-    Drawer,
-    DropdownDivider,
-    Sidebar,
-    SidebarGroup,
-    SidebarItem,
-    SidebarWrapper
-  } from 'flowbite-svelte'
-  import { CogSolid, HomeSolid, InfoCircleOutline } from 'flowbite-svelte-icons'
-  import { sineIn } from 'svelte/easing'
-  import { drawerHidden } from './stores.svelte'
+  import { HomeSolid, InfoCircleSolid } from 'flowbite-svelte-icons'
+  import { Drawer, Drawerhead, SidebarGroup, SidebarItem } from 'svelte-5-ui-lib'
+  import { useDrawerStore } from './stores.svelte'
 
-  const transitionParams = {
-    x: -320,
-    duration: 200,
-    easing: sineIn
-  }
+  const { drawerStatus, closeDrawer } = useDrawerStore()
 
   onNavigate(() => {
-    drawerHidden.value = true
+    closeDrawer()
   })
 
-  const activeUrl = $derived($page.url.pathname)
+  const navIconClass =
+    'h-5 w-5 text-gray-500 transition duration-75 group-hover:text-gray-900 dark:text-gray-400 dark:group-hover:text-white'
 </script>
 
-<Drawer transitionType="fly" bind:hidden={drawerHidden.value} {transitionParams}>
-  <div class="flex items-center">
+<Drawer drawerStatus={drawerStatus.value} {closeDrawer}>
+  <Drawerhead onclick={closeDrawer} class="mb-4">
     <h5
-      id="drawer-navigation-label-3"
-      class="text-base font-semibold text-gray-500 uppercase dark:text-gray-400"
+      id="drawer-label"
+      class="inline-flex items-center text-xl font-semibold text-gray-500 dark:text-gray-400"
     >
-      Menu
+      <InfoCircleSolid class="me-2.5 h-5 w-5" />Menu
     </h5>
-    <CloseButton onclick={() => (drawerHidden.value = true)} class="mb-4 dark:text-white" />
-  </div>
+  </Drawerhead>
 
-  <Sidebar {activeUrl}>
-    <SidebarWrapper divClass="overflow-y-auto py-4 px-3 rounded dark:bg-gray-800">
-      <SidebarGroup>
-        {@const navIconClass = `w-5 h-5 text-gray-500 transition duration-75
-          dark:text-gray-400group-hover:text-gray-900 dark:group-hover:text-white`}
-
-        <SidebarItem label="Home" href="/">
-          <svelte:fragment slot="icon">
-            <HomeSolid class={navIconClass} />
-          </svelte:fragment>
-        </SidebarItem>
-        <SidebarItem label="About" href="/about">
-          <svelte:fragment slot="icon">
-            <InfoCircleOutline class={navIconClass} />
-          </svelte:fragment>
-        </SidebarItem>
-
-        <DropdownDivider />
-
-        <SidebarItem label="Setting" href="/settings">
-          <svelte:fragment slot="icon">
-            <CogSolid class={navIconClass} />
-          </svelte:fragment>
-        </SidebarItem>
-      </SidebarGroup>
-    </SidebarWrapper>
-  </Sidebar>
+  <SidebarGroup>
+    <SidebarItem label="Home" href="/">
+      {#snippet iconSlot()}
+        <HomeSolid class={navIconClass} />
+      {/snippet}
+    </SidebarItem>
+    <SidebarItem label="About" href="/about">
+      {#snippet iconSlot()}
+        <InfoCircleSolid class={navIconClass} />
+      {/snippet}
+    </SidebarItem>
+  </SidebarGroup>
 </Drawer>
