@@ -1,18 +1,15 @@
-/** @type {import('tailwindcss').Config} */
+import type { Config } from 'tailwindcss'
+
 export default {
   content: [
     './src/**/*.{html,js,svelte,ts}',
-    './node_modules/flowbite-svelte/**/*.{html,js,svelte,ts}'
+    './node_modules/svelte-5-ui-lib/**/*.{html,js,svelte,ts}',
+    './node_modules/flowbite-svelte-icons/**/*.{html,js,svelte,ts}'
   ],
-
-  plugins: [require('flowbite/plugin')],
-
   darkMode: 'selector',
-
   theme: {
     extend: {
       colors: {
-        // flowbite-svelte
         primary: {
           50: '#FFF5F2',
           100: '#FFF1EE',
@@ -24,8 +21,20 @@ export default {
           700: '#EB4F27',
           800: '#CC4522',
           900: '#A5371B'
+        },
+        secondary: {
+          '50': '#f0f9ff',
+          '100': '#e0f2fe',
+          '200': '#bae6fd',
+          '300': '#7dd3fc',
+          '400': '#38bdf8',
+          '500': '#0ea5e9',
+          '600': '#0284c7',
+          '700': '#0369a1',
+          '800': '#075985',
+          '900': '#0c4a6e'
         }
       }
     }
   }
-}
+} as Config
